Create separate component for student and mentor cards
